+++ frontend/src/app/login/login.css
@@ -1,0 +1,99 @@
+/* Login.css */
+
+body,
+html {
+  height: 100%;
+  margin: 0;
+}
+
+.login-container {
+  position: relative;
+  height: 100%;
+}
+
+.login-box {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 15px;
+  border-radius: 12px;
+  background: #fff;
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
+  width: 70%;
+  max-width: 500px;
+}
+
+.logo-container {
+  display: inline-flex;
+  padding: 15px;
+  align-items: flex-start;
+  gap: 10px;
+  border-radius: 12px;
+  background: #fff;
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
+}
+
+.logo-image {
+  background: lightgray no-repeat;
+  width: 190px;
+  height: 90px;
+}
+
+.welcome-text {
+  color: #000;
+  text-align: center;
+  font-family: Lora;
+  font-size: 40px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: normal;
+  margin-top: 50px; /* Gap between Welcome and Name field */
+}
+
+.login-form {
+  width: 100%;
+  margin-top: 22px; /* Gap between Name and Password */
+}
+
+.input-group {
+  margin-bottom: 22px; /* Gap between Password and Forgot Password */
+}
+
+label {
+  display: block;
+  margin-bottom: 5px;
+  color: gray;
+}
+
+input {
+  width: 100%;
+  padding: 8px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+}
+
+.forgot-password {
+  color: blue;
+  text-align: center;
+  margin-bottom: 48px; /* Gap between Forgot password and Log In */
+}
+
+.login-button {
+  width: 100%;
+  height: 50px;
+  flex-shrink: 0;
+  border-radius: 4px;
+  background: var(--Secondary-1, #102d5f);
+  color: #fff;
+  cursor: pointer;
+  text-align: center;
+  font-family: Lora;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: normal;
+}
