--- frontend/src/app/login/page.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-export default function LoginPage() {
-  return (
-    <>
-      <main>
-        <div>{"Login page!"}</div>
-      </main>
-    </>
-  );
-}
